+++ devolo_home_control_api/backend/mprm_rest.py
@@ -80,3 +80,10 @@ class MprmRest:
             self._logger.error("Got an unexpected response after posting data.")
             raise ValueError("Got an unexpected response after posting data.")
         return response
+
+    def refresh_session(self):
+        """ Refresh curretly running session. Without this call from time to time, especially websockets will terminate. """
+        self._logger.debug("Refreshing session.")
+        data = {"method": "FIM/invokeOperation",
+                "params": [f"devolo.UserPrefs.{self._mydevolo.uuid()}", "resetSessionTimeout", []]}
+        self.post(data)