+++ devolo_home_control_api/properties/binary_sensor_property.py
@@ -20,7 +20,8 @@ class BinarySensorProperty(SensorProperty):
     """
 
     def __init__(self, gateway: Gateway, session: Session, element_uid: str, **kwargs: Any):
-        if not element_uid.startswith("devolo.BinarySensor:"):
+        if not element_uid.startswith(("devolo.BinarySensor:",
+                                       "devolo.SirenBinarySensor:")):
             raise WrongElementError(f"{element_uid} is not a Binary Sensor.")
 
         self.state = kwargs.get("state")
@@ -38,4 +39,5 @@ class BinarySensorProperty(SensorProperty):
     @last_activity.setter
     def last_activity(self, timestamp: int):
         """ Convert a timestamp in millisecond to a datetime object. """
-        self._last_activity = datetime.fromtimestamp(timestamp / 1000)
+        if timestamp != -1:
+            self._last_activity = datetime.fromtimestamp(timestamp / 1000)