+++ tests/mocks/mock_websocket.py
@@ -16,3 +16,7 @@ class MockWebsocket(MprmRest):
 class MockWebsocketError:
     def close(self):
         raise AssertionError
+
+
+def try_reconnect(self, sleep_interval):
+    self._reachable = True