+++ tests/test_mprm_websocket.py
@@ -1,7 +1,9 @@
 import time
-import requests
 
 import pytest
+
+from devolo_home_control_api.backend.mprm_rest import MprmRest
+from devolo_home_control_api.backend.mprm_websocket import MprmWebsocket
 
 from .mocks.mock_websocket import MockWebsocket, MockWebsocketError
 
@@ -27,6 +29,12 @@ class TestMprmWebsocket:
         with pytest.raises(AssertionError):
             self.mprm.websocket_disconnect()
 
+    @pytest.mark.usefixtures("mock_mprmwebsocket_websocket_disconnect")
+    def test__exit__(self, mocker):
+        disconnect_spy = mocker.spy(MprmWebsocket, "websocket_disconnect")
+        self.mprm.__exit__(None, None, None)
+        assert disconnect_spy.call_count == 1
+
     def test__on_message(self):
         with pytest.raises(NotImplementedError):
             message = '{"properties": {"com.prosyst.mbs.services.remote.event.sequence.number": 0}}'
@@ -39,17 +47,23 @@ class TestMprmWebsocket:
         assert event_sequence != self.mprm._event_sequence
         assert self.mprm._event_sequence == 5
 
-    # TODO: Fix test for MprmWebsocket._on_error
-    # @pytest.mark.usefixtures("mock_mprmwebsocket_get_remote_session")
-    # @pytest.mark.usefixtures("mock_mprmwebsocket_websocket_connection")
-    # def test__on_error(self):
-    #     self.mprm._ws = MockWebsocket()
-    #     self.mprm._on_error("error")
+    @pytest.mark.usefixtures("mock_mprmwebsocket_get_remote_session")
+    @pytest.mark.usefixtures("mock_mprmwebsocket_websocket_connection")
+    @pytest.mark.usefixtures("mock_mprmwebsocket_try_reconnect")
+    def test__on_error(self, mocker):
+        close_spy = mocker.spy(MockWebsocket, "close")
+        reconnect_spy = mocker.spy(MprmWebsocket, "_try_reconnect")
+        connect_spy = mocker.spy(MprmWebsocket, "websocket_connect")
+        self.mprm._ws = MockWebsocket()
+        self.mprm._on_error("error")
+        assert close_spy.call_count == 1
+        assert reconnect_spy.call_count == 1
+        assert connect_spy.call_count == 1
 
     @pytest.mark.usefixtures("mock_mprmwebsocket_websocketapp")
     @pytest.mark.usefixtures("mock_session_get")
     def test__on_pong(self, mocker, mprm_session, gateway_instance):
-        spy = mocker.spy(requests.Session, "get")
+        spy = mocker.spy(MprmRest, "refresh_session")
         self.mprm._session = mprm_session
         self.mprm.gateway = gateway_instance
         self.mprm._local_ip = self.gateway.get("local_ip")
@@ -57,7 +71,7 @@ class TestMprmWebsocket:
         assert spy.call_count == 1
 
     @pytest.mark.usefixtures("mock_mprmwebsocket_get_local_session_json_decode_error")
-    def test__try_reconnect(self, mocker, ):
+    def test__try_reconnect(self, mocker):
         spy = mocker.spy(time, "sleep")
         self.mprm._ws = MockWebsocket()
         self.mprm._local_ip = self.gateway.get("local_ip")